fix: moved all-clothes file to reducer folder and fixed dead links
