+++ client/components/navbar.js
@@ -22,7 +22,7 @@ const Navbar = ({handleClick, isLoggedIn}) => (
           {/* The navbar will show these links before you log in */}
           <Link to="/login">Login</Link>
           <Link to="/signup">Sign Up</Link>
-          <Link to="clothes">Clothes</Link>
+          <Link to="/clothes">Clothes</Link>
         </div>
       )}
     </nav>
