+++ client/reducers/all-clothes.js
