+++ client/store/index.js
@@ -2,13 +2,12 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import {clothes} from './all-clothes'
+// import user from './user'
+import {clothes} from '../reducers/all-clothes'
 import user from '../reducers/user'
 import item from '../reducers/item'
 
 const reducer = combineReducers({user, item, clothes})
-
 
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
